created AuthController and the get, post methods
get will bring the normal views, post will do the validations.

The Auth Facade that Laravel provides has several methods that you can
call on it. One of such is the attempt method. It takes in two
arguments. The first argument is an array of the fields you want to do
actual validation of against the database. In the background, it simply
queries the database with the email and password values to check if any
other user already has those details. If it’s present, it returns false,
if not it returns true. The second argument is the remember me
functionality. If the remember me checkbox was ticked in the form, then
it passes a value of true to the attempt method, otherwise false. When
the user clicks login, if the remember me checkbox was ticked, Laravel
creates a session cookie in the browser for that particular user. If the
user leaves the application without logging out, when the user returns,
Laravel just validates if the cookie is still there, if it is…then the
user won’t have to login again.